merge

+++ src/Chash/Command/User/DisableAdminsCommand.php
@@ -2,7 +2,7 @@
 
 namespace Chash\Command\User;
 
-use Symfony\Component\Console\Command\Command;
+use Chash\Command\Database\CommonDatabaseCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -14,7 +14,7 @@ use Symfony\Component\Console\Question\ConfirmationQuestion;
  * Remove the "admin" role from *ALL* users on all portals of this instance
  * @package Chash\Command\User
  */
-class DisableAdminsCommand extends CommonChamiloUserCommand
+class DisableAdminsCommand extends CommonDatabaseCommand
 {
     /**
      *
@@ -37,8 +37,8 @@ class DisableAdminsCommand extends CommonChamiloUserCommand
     {
         parent::execute($input, $output);
         $_configuration = $this->getHelper('configuration')->getConfiguration();
-        $connection = $this->getConnection($input);
         $helper = $this->getHelperSet()->get('question');
+        $conn = $this->getConnection($input);
         $question = new ConfirmationQuestion(
             '<question>This action will make all admins normal teachers. Are you sure? (y/N)</question>',
             false
@@ -47,12 +47,17 @@ class DisableAdminsCommand extends CommonChamiloUserCommand
             return;
         }
 
-        $us = "DELETE FROM admin";
-        $uq = mysql_query($us);
-        if ($uq === false) {
-            $output->writeln('Could not delete admins.');
-        } else {
-            $output->writeln('All admins disabled.');
+        if ($conn instanceof \Doctrine\DBAL\Connection) {
+            try {
+                $us = "DELETE FROM admin";
+                $stmt = $conn->prepare($us);
+                $stmt->execute();
+            } catch (\PDOException $e) {
+                $output->writeln('Could not delete admins.');
+                $output->write('SQL error!'.PHP_EOL);
+                throw new \RuntimeException($e->getMessage(), $e->getCode(), $e);
+            }
+            $output->writeln('All admins have been disabled. Use user:make-admin to add one back.');
         }
         return null;
     }
